+++ tests/mapWalk.test.js
@@ -67,6 +67,7 @@ async function mountPlains() {
       onMove: (position) => moves.push({ ...position }),
       onAction: () => {},
       onEncounter: () => {},
+      onWornPathHint: () => {},
       onCacheFound: () => {},
       onEnterMiniDungeon: () => {},
       onLockedGate: () => {},